+++ pranama-ui/src/components/View/services/email_service/ContactForm.vue
@@ -123,7 +123,7 @@
   </div>
 </template>
 <script>
-import ContactService from "@/shared/api/ContactService";
+import ContactService from "@/apiservices/ContactService";
 
 export default {
   data() {
